+++ docs/INVENTORY-MANIFEST.json
@@ -378,6 +378,7 @@
       "graphify.cjs",
       "gsd2-import.cjs",
       "handshake-serialized.cjs",
+      "health-diagnostic-types.cjs",
       "health-diagnostic.cjs",
       "hook-bus.cjs",
       "host-integration-sdk.cjs",
