+++ src/health-diagnostic-rules/config-validation.cts
@@ -54,8 +54,8 @@ import type planningSnapshotMod = require('../planning-snapshot.cjs');
 
 type PlanningSnapshot = ReturnType<typeof planningSnapshotMod.buildPlanningSnapshot>;
 
-// eslint-disable-next-line @typescript-eslint/no-require-imports -- health-diagnostic.cjs is an export= CommonJS module
-import healthDiagnosticMod = require('../health-diagnostic.cjs');
+// eslint-disable-next-line @typescript-eslint/no-require-imports -- health-diagnostic-types.cjs is an export= CommonJS module
+import healthDiagnosticMod = require('../health-diagnostic-types.cjs');
 const { SEVERITY, REMEDY_ACTION, REMEDY_RISK } = healthDiagnosticMod;
 type Diagnostic = healthDiagnosticMod.Diagnostic;
 type Remedy = healthDiagnosticMod.Remedy;
